# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "pubchem api server",
        "title": "Golang go-pubchem APIs",
        "termsOfService": "https://www.swagger.io/terms/",
        "contact": {
            "name": "chengxiangLuo",
            "url": "https://github.com/netluo",
            "email": "andrew.luo1992@gmile.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.21.94:8100",
    "basePath": "/api/v1",
    "paths": {
        "/db/addDefaultValue": {
            "post": {
                "description": "insert compound info to db by cid",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "db"
                ],
                "summary": "AddDefaultValueForVarchar 设置表的字段为非空，并设置默认值",
                "parameters": [
                    {
                        "description": "Cid",
                        "name": "cid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src.DbAndTableName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"hello wy\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"who are you\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/db/fillNullValue": {
            "post": {
                "description": "insert compound info to db by cid",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "db"
                ],
                "summary": "FillNullValue 填充表的 null 值",
                "parameters": [
                    {
                        "description": "Cid",
                        "name": "cid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/src.DbAndTableName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"hello wy\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"who are you\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "src.DbAndTableName": {
            "type": "object",
            "properties": {
                "dbName": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        }
    }
}